mach-snapdragon: add board_spl.c and split out common code
Code in board.c will now only be compiled into U-Boot proper,
and the new board_spl.c will only be built into SPL.
Code in mem_map.c is common to both phases since it seems to not cause
issues in SPL. The existing dram.c is also common to both phases with
similar reasoning.
In the future memory map related code should probably behave differenly
in SPL, especially if dram initialization is supported.
